Add fuel consumption to static data.
Fuel consumption is a dictionary keyed on fuel type, where the values are
dictionaries with keys "perhour", "purpose", and maybe "empire". The "perhour"
value is the amount of that resource consumed per hour, the "purpose" value is
the use of that resource (currently only "online" or "reinforce"), and the
"empire" value, if present, is the name of the NPC faction in whose highsec
space this resource is consumed. The "empire" key is currently used only for
charters.